refactor: slipt util/index.ts into multiple files
