+++ packages/web-extension/src/pages/SessionList.tsx
@@ -31,7 +31,7 @@ import { VscTriangleDown, VscTriangleUp } from 'react-icons/vsc';
 import { useNavigate } from 'react-router-dom';
 import { Session, EventName } from '~/types';
 import Channel from '~/utils/channel';
-import { deleteSessions, getAllSessions } from '~/utils';
+import { deleteSessions, getAllSessions } from '~/utils/storage';
 import {
   FiChevronLeft,
   FiChevronRight,
